Calling the Microsoft Cognitive Services API. Notes: This is just the sample code, modified to work on Android. A few notes:
Permissions must be set in the Manifest to allow Internet Access.
Network calls, if done on the Main thread, throw an exception and fail.
The apache libraries are deprecated, but can still be used by adding the changes I made to the .gradle file.

(This code is in no way representative of the way I'd recommend people to write their production code)